refactor producer
